Refactored completely , and added functionality to export to excel.
The new refactoring allows for a more modular relating between components. now in 2 phases , export, for which there are specific classes, and the other is format. 

The new functionality to export to excel is just a new class to format in excel , where formerly there was only option to export in csv. 

There is a new sample configuration file in Inmersa-cron.ini.sample that will relate to changes in configuration variable names. 

Tite